Use the SwResultExt trait in more places
